imr/alloc: extract lsa specific code to logalloc.hh
The lsa specific code is not used in the alloc.hh, only by users, but it
pulls in imr/methods.hh and indirectly imr/compound.hh. We want to
include imr/alloc.hh in imr/compound.hh to be able to spell out the type
of the allocators used in sizing and serializing but that would
introduce a circular dependency. To solve this we extract the lsa
specific code into a separate header, which users like imr/utils.hh can
include.
